Refactor ARC-related code: remove unused files, implement ARC debug firmware interactions, and enhance ARC core control methods
