--- ttexalens/hw/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
-
-# SPDX-License-Identifier: Apache-2.0
